# Symend.Server.Customer/src/IO.Swagger/Models/CustomerTrailEventTypeCategory.cs
/*
 * Customer Service API
 *
 * API for Customer Domain Operations
 *
 * OpenAPI spec version: v1
 * Contact: support@symend.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
        /// <summary>
        /// Gets or Sets CustomerTrailEventTypeCategory
        /// </summary>
        [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
        public enum CustomerTrailEventTypeCategory
        {
            /// <summary>
            /// Enum UnknownEnum for unknown
            /// </summary>
            [EnumMember(Value = "unknown")]
            UnknownEnum = 0,
            /// <summary>
            /// Enum AccountUpdatesEnum for accountUpdates
            /// </summary>
            [EnumMember(Value = "accountUpdates")]
            AccountUpdatesEnum = 1,
            /// <summary>
            /// Enum EmailActivitiesEnum for emailActivities
            /// </summary>
            [EnumMember(Value = "emailActivities")]
            EmailActivitiesEnum = 2,
            /// <summary>
            /// Enum SmsActivitiesEnum for smsActivities
            /// </summary>
            [EnumMember(Value = "smsActivities")]
            SmsActivitiesEnum = 3,
            /// <summary>
            /// Enum PlaybookActivitiesEnum for playbookActivities
            /// </summary>
            [EnumMember(Value = "playbookActivities")]
            PlaybookActivitiesEnum = 4,
            /// <summary>
            /// Enum CampaignActivitiesEnum for campaignActivities
            /// </summary>
            [EnumMember(Value = "campaignActivities")]
            CampaignActivitiesEnum = 5,
            /// <summary>
            /// Enum LandingPageActivitiesEnum for landingPageActivities
            /// </summary>
            [EnumMember(Value = "landingPageActivities")]
            LandingPageActivitiesEnum = 6        }
}